// src/api/models/offlineTicketListItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { OfflineTicketListItemUser } from './offlineTicketListItemUser';
import type { OfflineTicketStoreItem } from './offlineTicketStoreItem';

export interface OfflineTicketListItem {
  /** ID */
  id: string;
  /** 현장권 상태 */
  status: string;
  /** 결제 수단 */
  paymentMethod: string;
  /** 서비스 종류 */
  serviceType: string;
  /**
   * 서비스 옵션
   * @nullable
   */
  serviceOptions?: string | null;
  /**
   * 차량 번호
   * @nullable
   */
  carNumber?: string | null;
  /** 생성일 */
  createdAt: string;
  /**
   * 회원 정보
   * @nullable
   */
  user: OfflineTicketListItemUser;
  /** 매장 정보 */
  store: OfflineTicketStoreItem;
}